+++ BackEnd/Areas/Identity/Data/TheMenuBackEndContext.cs
@@ -18,5 +18,10 @@ public class TheMenuBackEndContext : IdentityDbContext<TheMenuBackEndUser>
         // Customize the ASP.NET Identity model and override the defaults if needed.
         // For example, you can rename the ASP.NET Identity table names and more.
         // Add your customizations after calling base.OnModelCreating(builder);
+
+        //builder.Entity<User>().ToTable("AspNetUsers");
     }
+
+    //public DbSet<User>? Users { get; set; }
+ 
 }